refactor(krites/data): full structural overhaul — snafu errors, quality (#3214)
- Replace panic in `From<DataValue> for JsonValue` with `TryFrom` returning
  `DataValueToJsonError` for Bot sentinel — the only non-library panic in data/
- Split functions/aggregate.rs (730 LOC) into aggregate.rs (330) + collections.rs (442)
- Add `Debug` derives to UuidWrapper, RegexWrapper, JsonData, Validity
- Document all DataValue variants, Num, Vector, ValidityTs, Validity, JsonData,
  UuidWrapper, RegexWrapper, aggregation traits and contracts
- Fix Num doc typo ("intger" -> "integer" via full rewrite)
- Add module-level doc to data/mod.rs
- Fix pre-existing clippy: `Default::default()` -> `String::new()`, unused
  `mutable_key_type` expect, `serde_json` backticks in doc comment
- All callers updated: runtime/db.rs, relation.rs coercion, test/json.rs
- New test: `bot_to_json_returns_error` verifies TryFrom rejects Bot

Verified: cargo check -p krites, cargo test -p krites (329 pass),
cargo test -p episteme, cargo test -p mneme, cargo clippy -p krites (0 new warnings)
