+++ crates/krites/src/data/value.rs
@@ -32,7 +32,11 @@ use uuid::Uuid;
 use crate::data::json::JsonValue;
 use crate::data::relation::VecElementType;
 
-#[derive(Clone, Hash, Eq, PartialEq, serde::Deserialize, serde::Serialize)]
+/// Newtype wrapper around [`Uuid`] providing custom `Ord` for memcmp-compatible key ordering.
+///
+/// UUID fields are compared in (time_hi, time_mid, time_low, rest) order so that
+/// v1 UUIDs sort chronologically in the storage layer.
+#[derive(Clone, Debug, Hash, Eq, PartialEq, serde::Deserialize, serde::Serialize)]
 pub struct UuidWrapper(pub Uuid);
 
 impl PartialOrd<Self> for UuidWrapper {
@@ -52,9 +56,19 @@ impl Ord for UuidWrapper {
     }
 }
 
-/// A Regex in the database. Used internally in functions.
+/// Compiled regex carried as a transient value inside the engine.
+///
+/// `RegexWrapper` is **not** serializable: it exists only during expression
+/// evaluation (e.g., `regex_matches`). `Hash`, `Eq`, and `Ord` compare the
+/// source pattern string, not compiled state.
 #[derive(Clone)]
 pub struct RegexWrapper(pub Regex);
+
+impl Debug for RegexWrapper {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "Regex({:?})", self.0.as_str())
+    }
+}
 
 impl Hash for RegexWrapper {
     fn hash<H: Hasher>(&self, state: &mut H) {
@@ -104,20 +118,27 @@ impl PartialOrd for RegexWrapper {
     }
 }
 
-/// Timestamp part of validity
+/// Microsecond timestamp for time-travel validity, sorted **descending**.
+///
+/// Wraps `Reverse<i64>` so that the natural `Ord` puts newer timestamps first,
+/// which is required by the key-scan logic in `check_key_for_validity`.
 #[derive(
     Copy, Clone, Eq, PartialEq, Ord, PartialOrd, serde::Deserialize, serde::Serialize, Hash, Debug,
 )]
 pub struct ValidityTs(pub Reverse<i64>);
 
-/// Validity for time travel
+/// Validity marker attached to stored tuples for time-travel queries.
+///
+/// Each stored fact carries a `(timestamp, is_assert)` pair. Assertions add the
+/// fact at a point in time; retractions remove it. Both fields sort descending so
+/// that the storage scan finds the most recent state first.
 #[derive(
-    Copy, Clone, Eq, PartialEq, Ord, PartialOrd, serde::Deserialize, serde::Serialize, Hash,
+    Copy, Clone, Debug, Eq, PartialEq, Ord, PartialOrd, serde::Deserialize, serde::Serialize, Hash,
 )]
 pub struct Validity {
-    /// Timestamp, sorted descendingly
+    /// Microsecond timestamp, sorted descending (newest first).
     pub timestamp: ValidityTs,
-    /// Whether this validity is an assertion, sorted descendingly
+    /// `true` = assertion, `false` = retraction; sorted descending.
     pub is_assert: Reverse<bool>,
 }
 
@@ -130,42 +151,63 @@ impl From<(i64, bool)> for Validity {
     }
 }
 
-/// A Value in the database
+/// The core value type for every datum in the Datalog engine.
+///
+/// `DataValue` appears in tuples, expressions, function arguments, and storage
+/// keys. Variant ordering defines the memcmp sort order used by the storage
+/// layer (Null < Bool < Num < Str < ... < Bot).
+///
+/// # Internal-only variants
+///
+/// `Regex`, `Set`, `Validity`, and `Bot` are engine-internal: they never appear
+/// in user-facing query results. `Bot` is the bottom sentinel used as the
+/// upper-bound key in range scans.
 #[derive(Clone, PartialEq, Eq, PartialOrd, Ord, serde::Deserialize, serde::Serialize, Hash)]
 #[non_exhaustive]
 pub enum DataValue {
-    /// null
+    /// The null (absent) value. Sorts before all other variants.
     Null,
-    /// boolean
+    /// Boolean truth value.
     Bool(bool),
-    /// number, may be int or float
+    /// Numeric value — integer or float (see [`Num`]).
     Num(Num),
-    /// string
+    /// UTF-8 string, stored inline via [`CompactString`].
     Str(CompactString),
-    /// bytes
+    /// Raw byte sequence, serialized via `serde_bytes`.
     #[serde(with = "serde_bytes")]
     Bytes(Vec<u8>),
-    /// UUID
+    /// UUID value with chronological sort order (see [`UuidWrapper`]).
     Uuid(UuidWrapper),
-    /// Regex, used internally only
+    /// Compiled regex — transient, not serializable. Engine-internal.
     Regex(RegexWrapper),
-    /// list
+    /// Ordered sequence of values.
     List(Vec<DataValue>),
-    /// set, used internally only
+    /// Deduplicated ordered set. Engine-internal; coerced to `List` at output.
     Set(BTreeSet<DataValue>),
-    /// Array, mainly for proximity search
+    /// Typed floating-point vector for proximity search (HNSW).
     Vec(Vector),
-    /// Json
+    /// Arbitrary JSON value (objects, arrays, etc.).
     Json(JsonData),
-    /// validity,
+    /// Timestamp + assertion flag for time-travel queries. Engine-internal.
     Validity(Validity),
-    /// bottom type, used internally only
+    /// Bottom sentinel — sorts after everything. Used as upper key bound. Engine-internal.
     Bot,
 }
 
-/// Wrapper for JsonValue
+/// Wrapper around [`JsonValue`] that provides `Ord` and `Hash` (via string representation).
+///
+/// JSON objects and arrays are stored as opaque blobs in the Datalog engine.
+/// Ordering and hashing use the JSON serialized form, which is sufficient for
+/// key deduplication but **not** semantically meaningful (different key orders
+/// in objects compare differently).
 #[derive(Clone, PartialEq, Eq, serde::Deserialize, serde::Serialize)]
 pub struct JsonData(pub JsonValue);
+
+impl Debug for JsonData {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        write!(f, "JsonData({})", self.0)
+    }
+}
 
 impl PartialOrd<Self> for JsonData {
     fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
@@ -193,13 +235,17 @@ impl Hash for JsonData {
     }
 }
 
-/// Vector of floating numbers
+/// Typed dense vector of floating-point numbers, used for HNSW proximity search.
+///
+/// Supports both single- and double-precision representations. Distance
+/// functions (`l2_dist`, `ip_dist`, `cos_dist`) operate on vectors of the
+/// same element type.
 #[derive(Debug, Clone)]
 #[non_exhaustive]
 pub enum Vector {
-    /// 32-bit float array
+    /// Single-precision (32-bit) float array.
     F32(Array1<f32>),
-    /// 64-bit float array
+    /// Double-precision (64-bit) float array.
     F64(Array1<f64>),
 }
 
@@ -479,13 +525,17 @@ impl From<bool> for DataValue {
     }
 }
 
-/// Representing a number
+/// Numeric value: either an exact integer or an IEEE 754 double.
+///
+/// Mixed-type arithmetic promotes `Int` to `Float`. The custom `Ord`
+/// implementation sorts `Int(n)` before `Float(n)` when the float
+/// representation is identical, preserving type distinction in key ordering.
 #[derive(Copy, Clone, serde::Deserialize, serde::Serialize)]
 #[non_exhaustive]
 pub enum Num {
-    /// intger number
+    /// Exact integer value.
     Int(i64),
-    /// float number
+    /// IEEE 754 double-precision floating-point value.
     Float(f64),
 }
 
@@ -520,6 +570,7 @@ impl Num {
             }
         }
     }
+    /// Convert to f64, promoting integers via `as` cast (precision loss above 2^53).
     pub(crate) fn get_float(&self) -> f64 {
         match self {
             #[expect(
@@ -689,6 +740,7 @@ impl DataValue {
             _ => None,
         }
     }
+    /// Returns the integer value if non-negative, converting whole floats.
     pub(crate) fn get_non_neg_int(&self) -> Option<u64> {
         match self {
             DataValue::Num(n) => n.get_int().and_then(|i| u64::try_from(i).ok()),
@@ -709,9 +761,11 @@ impl DataValue {
             _ => None,
         }
     }
+    /// Construct a `DataValue::Uuid` from a raw [`Uuid`].
     pub(crate) fn uuid(uuid: Uuid) -> Self {
         Self::Uuid(UuidWrapper(uuid))
     }
+    /// Extract UUID, also parsing from `Str` if the string is a valid UUID.
     pub(crate) fn get_uuid(&self) -> Option<Uuid> {
         match self {
             DataValue::Uuid(UuidWrapper(uuid)) => Some(*uuid),
@@ -721,4 +775,5 @@ impl DataValue {
     }
 }
 
+/// Largest valid Unicode code point — used as the suffix for prefix-scan upper bounds.
 pub(crate) const LARGEST_UTF_CHAR: char = '\u{10ffff}';
